requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
debugpy==1.5.1
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.0.0
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.2
gunicorn==20.1.0
idna==3.3
jedi==0.18.1
mypy-extensions==0.4.3
nest-asyncio==1.5.4
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
whitenoise==6.0.0
